pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool]
[tool.poetry]
authors = ["Your Name <you@example.com>"]
description = ""
name = "python-template"
version = "0.1.0"
[tool.poetry.dependencies]
numpy = "^1.22.2"
python = "^3.8"
emoji = "^1.6.3"
py-cord = "^1.7.3"
emoji-data-python = "^1.5.0"
psutil = "^5.9.0"
pyfiglet = "^0.8.post1"
Pygments = "^2.11.2"
qrcode = "^7.3.1"
python-dotenv = "^0.19.2"
grequests = "^0.6.0"
aiohttp = "3.7.*"
[tool.poetry.dev-dependencies]
